requirements.txt
django>=5.0,<5.1
django-ninja>=1.3
celery>=5.4
flower>=2.0
redis>=5.0
psycopg[binary]>=3.1
agno>=2.5
anthropic>=0.40
